Add support for bootstrapping additional root apps
See [SDD-0030 - ArgoCD multitenancy] for the overall design for
additional root apps.

This commit introduces a new command line flag to specify a config map
from which Steward will read field `teams` which is expected to contain
a JSON-encoded list of team names for which to setup additional ArgoCD
AppProject and root Application (app of apps) resources.

Steward uses the team names as is for the additional AppProject
resources, and creates the additional root apps as `root-<team name>`.
As defined in the design document, the path for the root apps is
configured as `manifests/apps-<team name>/`.

The new code is added to `argocd.Apply()` before the short-circuit which
exits the function if an ArgoCD custom resource is present on the
cluster already.

[SDD-0030 - ArgoCD multitenancy]: https://syn.tools/syn/SDDs/0030-argocd-multitenancy.html